% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04ptwyunting-package.R
\docType{package}
\name{hw04ptwyunting-package}
\alias{hw04ptwyunting}
\alias{hw04ptwyunting-package}
\title{hw04ptwyunting: My first personal R package}
\description{
This package aims to make a math formula in sequence with a proper graph.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/STAT-413-613-Fall-2020/hw04ptwyunting}
  \item Report bugs at \url{https://github.com/STAT-413-613-Fall-2020/hw04ptwyunting/issues}
}

}
\author{
\strong{Maintainer}: Yunting Chiu \email{yc6705a@american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
