% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numSeqPlot.R
\name{numSeqPlot}
\alias{numSeqPlot}
\title{A line graph that plots the relation between first three columns and fourth column.}
\usage{
numSeqPlot(nums)
}
\arguments{
\item{nums}{A data frame with four columns, and each column must contain numbers only, the fourth column must be the positive integer.}
}
\value{
Return a line plot of the output values for the different values of n.
}
\description{
The prerequisite is this data frame must have four columns. The first three columns are the values of the three numeric to be input, and the fourth column is the positive integer n for the sequence to be generated.
This function should return a line plot of the output values for the different values of n.
}
\examples{
# testing the function as data frame
my_data <- tibble::tribble(
~x, ~y, ~z, ~n,
2,4,3,3,
2,4,3,4,
2,4,3,5,
2,4,3,6,
2,4,3,7,
2,4,3,8,
2,4,3,9,
2,4,3,10,
2,4,3,12)
numSeqPlot(my_data)
}
